requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
discord==1.0.1
discord.py==1.7.1
giphy-client==1.0.0
idna==2.10
multidict==5.1.0
musixmatch==0.9
pycparser==2.20
pymusixmatch==0.3
PyNaCl==1.4.0
pyrandmeme==0.0.6
python-dateutil==2.8.1
python-dotenv==0.19.2
random2==1.0.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
youtube-dl==2021.4.7
